requirements.txt
discord.py>=2.0.0 # Atau versi spesifik yang kamu gunakan
# Tambahkan library lain jika kamu menggunakannya, contoh:
# requests 